use payload type if no name
