+++ packages/typespec-ts/src/modular/helpers/operationHelpers.ts
@@ -150,7 +150,7 @@ export function getDeserializePrivateFunction(
         `if(${validStatus
           .map((s) => `result.status !== "${s}"`)
           .join(" || ")}){`,
-         `throw createRestError(result);`,
+        `throw createRestError(result);`,
         "}"
       );
       addImportToSpecifier("restClient", runtimeImports, "createRestError");
@@ -525,10 +525,7 @@ function getParameterMap(
   throw new Error(`Parameter ${param.clientName} is not supported`);
 }
 
-function getCollectionFormat(
-  param: Parameter,
-  runtimeImports: RuntimeImports
-) {
+function getCollectionFormat(param: Parameter, runtimeImports: RuntimeImports) {
   const collectionInfo = getCollectionFormatHelper(
     param.location,
     param.format ?? ""
@@ -590,10 +587,7 @@ function isRequired(param: Parameter | Property): param is RequiredType {
   return !param.optional;
 }
 
-function getRequired(
-  param: RequiredType,
-  runtimeImports: RuntimeImports
-) {
+function getRequired(param: RequiredType, runtimeImports: RuntimeImports) {
   if (param.type.type === "model") {
     return `"${param.restApiName}": {${getRequestModelMapping(
       param.type,
@@ -645,10 +639,7 @@ function isOptional(param: Parameter | Property): param is OptionalType {
   return Boolean(param.optional);
 }
 
-function getOptional(
-  param: OptionalType,
-  runtimeImports: RuntimeImports
-) {
+function getOptional(param: OptionalType, runtimeImports: RuntimeImports) {
   if (param.type.type === "model") {
     return `"${param.restApiName}": {${getRequestModelMapping(
       param.type,
@@ -938,7 +929,7 @@ function deserializeResponseValue(
     }
     case "byte-array":
       if (format !== "binary") {
-        addImportToSpecifier("coreUtil", runtimeImports, "stringToUint8Array"); 
+        addImportToSpecifier("coreUtil", runtimeImports, "stringToUint8Array");
         return `typeof ${restValue} === 'string'
         ? stringToUint8Array(${restValue}, "${format ?? "base64"}")
         : ${restValue}`;
